requirements.txt
celery~=5.2.7
requests==2.28.2
kafka-python~=2.0.2
yfinance~=0.2.12
pysingleton~=0.2.1